Task_4_Edge_VM/requirements.txt
numpy
matplotlib
paho-mqtt>=2.0.0
tensorflow
